Change dropdown class toggling
Instead of removing and adding classes, class toggling is enabled.

Dropdown menu doesn't show if triangle icon is clicked. This is due to
the window event listener closing the dropdown menu before it can be
opened. Fixed by adding a condition on the window event listener to
exclude the triangle icon from its logic.