Modified css code for slide indicator element
Read on stackoverflow of how to center an absolute positioned element.

https://stackoverflow.com/questions/1776915/how-can-i-center-an-absolutely-positioned-element-in-a-div

Left: 50% is used to make sure that the left corner of the element is
positioned at the middle of the parent.

Translate(-50%, 0) is then used to move the element halfway to the left.

This ensures responsiveness of the element to be at the middle whenever
the parent element's width changes.