Add function to change slide on indicator clicks
Changed setTimeout to setInterval as there were issues with setTimeout
stopping even after setting timeout again after clearTimeout

Refactored the showSlides() function, repurposing it to just load the
WITHOUT messing with the slideIndex variable.

setInterval function is now separated into the main body so that
eventListeners for the slide arrows and indicators can reference it
to reset.

Slide arrows and indicators is now able to change the slides by
modifying the slideIndex variable and utilizing the showSlides()
function after.

changeSlides() function is added to handle the usual logic of modifying
the slideIndex variable upon invokation. This function is used in the
setInterval function.